+++ apps/afisha/admin.py
@@ -1,12 +1,6 @@
 from django.contrib import admin
 
-from apps.afisha.models import CommonEvent, Event
-
-
-class CommonEventAdmin(admin.ModelAdmin):
-    list_display = ("pk", "target_model")
-    list_filter = ("created",)
-    empty_value_display = "-пусто-"
+from apps.afisha.models import Event
 
 
 class EventAdmin(admin.ModelAdmin):
@@ -32,4 +26,3 @@ class EventAdmin(admin.ModelAdmin):
 
 
 admin.site.register(Event, EventAdmin)
-admin.site.register(CommonEvent, CommonEventAdmin)
